+++ iOS/Controller/SideBar/BookmarkController.swift
@@ -1,6 +1,6 @@
 //
 //  BookmarkController.swift
-//  iOS
+//  Kiwix
 //
 //  Created by Chris Li on 5/21/18.
 //  Copyright © 2018 Chris Li. All rights reserved.
@@ -9,9 +9,8 @@
 import UIKit
 import RealmSwift
 
-class BookmarkController: UIViewController, UITableViewDataSource, UITableViewDelegate {
+class BookmarkController: UITableViewController {
     weak var delegate: BookmarkControllerDelegate? = nil
-    private let tableView = UITableView()
     private lazy var emptyContentView = EmptyContentView(
         image: #imageLiteral(resourceName: "StarColor"),
         title: NSLocalizedString("Bookmark your favorite articles", comment: "Help message when there's no bookmark to show"),
@@ -45,9 +44,9 @@ class BookmarkController: UIViewController, UITableViewDataSource, UITableViewDe
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        tableView.dataSource = self
-        tableView.delegate = self
-        tableView.register(TableViewCell.self, forCellReuseIdentifier: "Cell")
+        tableView.register(ArticleTableViewCell.self, forCellReuseIdentifier: "Cell")
+        tableView.separatorInsetReference = .fromAutomaticInsets
+        tableView.separatorInset = UIEdgeInsets(top: 0, left: 20, bottom: 0, right: 0)
         
         if let _ = presentingViewController {
             navigationItem.leftBarButtonItem = UIBarButtonItem(barButtonSystemItem: .done,
@@ -58,7 +57,7 @@ class BookmarkController: UIViewController, UITableViewDataSource, UITableViewDe
     
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
-        configureContentView()
+        configureBackground()
         configureChangeToken()
         tableView.reloadData()
     }
@@ -74,72 +73,65 @@ class BookmarkController: UIViewController, UITableViewDataSource, UITableViewDe
         dismiss(animated: true, completion: nil)
     }
     
-    private func configureContentView() {
-        if let bookmarks = bookmarks, bookmarks.count > 0 {
-            guard !view.subviews.contains(tableView) else {return}
-            view.subviews.forEach({ $0.removeFromSuperview() })
-            view.backgroundColor = .white
-            tableView.translatesAutoresizingMaskIntoConstraints = false
-            view.addSubview(tableView)
-            NSLayoutConstraint.activate([
-                view.topAnchor.constraint(equalTo: tableView.topAnchor),
-                view.leftAnchor.constraint(equalTo: tableView.leftAnchor),
-                view.bottomAnchor.constraint(equalTo: tableView.bottomAnchor),
-                view.rightAnchor.constraint(equalTo: tableView.rightAnchor)])
+    private func configureBackground() {
+        if bookmarks?.count ?? 0 > 0 {
+            tableView.backgroundView = nil
+            tableView.separatorStyle = .singleLine
         } else {
-            guard !view.subviews.contains(emptyContentView) else {return}
-            view.subviews.forEach({ $0.removeFromSuperview() })
-            view.backgroundColor = .groupTableViewBackground
-            emptyContentView.translatesAutoresizingMaskIntoConstraints = false
-            view.addSubview(emptyContentView)
-            NSLayoutConstraint.activate([
-                view.centerYAnchor.constraint(equalTo: emptyContentView.centerYAnchor),
-                view.leftAnchor.constraint(equalTo: emptyContentView.leftAnchor),
-                view.rightAnchor.constraint(equalTo: emptyContentView.rightAnchor)])
+            tableView.backgroundView = emptyContentView
+            tableView.separatorStyle = .none
         }
     }
     
     private func configureChangeToken() {
         changeToken = bookmarks?.observe({ (changes) in
             guard case .update(_, let deletions, let insertions, let updates) = changes else {return}
-            
-            self.tableView.beginUpdates()
-            self.tableView.deleteRows(at: deletions.map({ IndexPath(row: $0, section: 0) }), with: .fade)
-            self.tableView.insertRows(at: insertions.map({ IndexPath(row: $0, section: 0) }), with: .fade)
-            updates.forEach({ row in
-                let indexPath = IndexPath(row: row, section: 0)
-                guard let cell = self.tableView.cellForRow(at: indexPath) as? TableViewCell else {return}
-                self.configure(cell: cell, indexPath: indexPath)
-            })
-            self.tableView.endUpdates()
+            self.configureBackground()
+            self.tableView.performBatchUpdates({
+                self.configureBackground()
+                self.tableView.deleteRows(at: deletions.map({ IndexPath(row: $0, section: 0) }), with: .fade)
+                self.tableView.insertRows(at: insertions.map({ IndexPath(row: $0, section: 0) }), with: .fade)
+                updates.forEach({ row in
+                    let indexPath = IndexPath(row: row, section: 0)
+                    guard let cell = self.tableView.cellForRow(at: indexPath) as? TableViewCell else {return}
+                    self.configure(cell: cell, indexPath: indexPath)
+                })
+            }, completion: nil)
         })
     }
     
     // MARK: - UITableViewDataSource & Delagate
     
-    func numberOfSections(in tableView: UITableView) -> Int {
+    override func numberOfSections(in tableView: UITableView) -> Int {
         return 1
     }
 
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+    override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return bookmarks?.count ?? 0
     }
 
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+    override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "Cell", for: indexPath) as! TableViewCell
         configure(cell: cell, indexPath: indexPath)
         return cell
     }
 
     func configure(cell: TableViewCell, indexPath: IndexPath, animated: Bool = false) {
-        guard let bookmark = bookmarks?[indexPath.row] else {return}
+        guard let bookmark = bookmarks?[indexPath.row], let zimFile = bookmark.zimFile else {return}
         cell.titleLabel.text = bookmark.title
+        cell.titleLabel.font = UIFont.systemFont(ofSize: 16, weight: .medium)
         cell.detailLabel.text = bookmark.snippet
-        cell.thumbImageView.image = UIImage(data: bookmark.thumbImageData ?? Data()) ?? #imageLiteral(resourceName: "GenericZimFile")
-        cell.thumbImageView.contentMode = .scaleAspectFill
+        if let thumbImagePath = bookmark.thumbImagePath,
+           let data = ZimMultiReader.shared.getContent(bookID: zimFile.id, contentPath: thumbImagePath)?.data {
+            cell.thumbImageView.image = UIImage(data: data)
+        } else if let data = zimFile.faviconData {
+            cell.thumbImageView.image = UIImage(data: data)
+        } else {
+            cell.thumbImageView.image = #imageLiteral(resourceName: "GenericZimFile")
+        }
     }
     
-    func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
+    override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         guard let bookmark = bookmarks?[indexPath.row], let zimFileID = bookmark.zimFile?.id,
             let url = URL(zimFileID: zimFileID, contentPath: bookmark.path) else {return}
         tableView.deselectRow(at: indexPath, animated: true)
@@ -149,21 +141,14 @@ class BookmarkController: UIViewController, UITableViewDataSource, UITableViewDe
         }
     }
 
-    func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        guard let bookmark = bookmarks?[indexPath.row] else {return}
-        let url: URL? = {
-            guard let zimFileID = bookmark.zimFile?.id else {return nil}
-            return URL(zimFileID: zimFileID, contentPath: bookmark.path)
-        }()
-        if editingStyle == .delete {
-            do {
-                let database = try Realm(configuration: Realm.defaultConfig)
-                try database.write {
-                    database.delete(bookmark)
-                }
-                if let url = url { delegate?.didDeleteBookmark(url: url) }
-            } catch {}
+    override func tableView(_ tableView: UITableView,
+                            commit editingStyle: UITableViewCell.EditingStyle,
+                            forRowAt indexPath: IndexPath) {
+        guard let bookmark = bookmarks?[indexPath.row], editingStyle == .delete else {return}
+        if let zimFileID = bookmark.zimFile?.id, let url = URL(zimFileID: zimFileID, contentPath: bookmark.path) {
+            delegate?.didDeleteBookmark(url: url)
         }
+        BookmarkService().delete(bookmark)
     }
 }